pneumonia-prediction-master/requirements.txt
flask
tensorflow
Pillow>=8.0.0